dshow: fix warning on win32

+++ modules/dshow/dshow.cpp
@@ -146,7 +146,7 @@ public:
 		uint32_t tmp_pix_RGB32;
 		uint32_t *buf_RGB32;
 		struct vidframe vidframe;
-		uint64_t timestamp = sample_time * VIDEO_TIMEBASE;
+		uint64_t timestamp = (uint64_t)(sample_time * VIDEO_TIMEBASE);
 
 		vidframe_init_buf(&vidframe, VID_FMT_RGB32, &src->size, buf);
